Merged in feature/1-add-project-skeleton (pull request #1)
* requirements.txt uses shared library master branch

The required changes have been merged to master on shared library. Use
master branch to install the shared library for development.

* Use BBCREDS in Jenkinsfile & requirements.txt

The CDCAGG shared library is not a public repo and needs BitBucket
credentials to install via pip. Jenkinsfile provides the credentials
in an environment variable and requirements.txt uses the variable to
supply the credentials to the clone URL.

* Add templates to MANIFEST.in so they'll be packaged correctly.

* Add metadataformats entry points to setup.py

* Add project skeleton with Jenkinsfile & sonar-project.properties

Fixes #1