Fix empty setName elements for ListSets-verb
Empty setName elements for language-sets are populated by generated
values. This comes from dependency kuha_oai_pmh_repo_handler and is
implemented in e89c58787075e9e79dfb4a31b15d8b1129b1efcf, which is now
the changeset required by requirements.txt.

Mapping file syntax for source-sets now support setname and
description key-value pairs. The setname is mandatory and description
is optional. These values are used to populate setName and description
elements in response to ListSets-verb.

Mapping file configuration option for
SourceAggMDSet (--oai-set-sources-path) defaults to None, which
implies that the set is discarded on server startup. The configuration
and creation of the mapping file is left to the operator.

The file sources_default.yaml is renamed to sources_set.yaml.example
and is explicitly an example of a mapping file.

Dockerfile no longer copies the sources_default.yaml to include it in the
container image.

Source set (SourceAggMDSet) no longer uses source archives baseUrl as
a fall back value for setSpec.

Write items to changelog entry for 0.2.0.

Issue #15 at Bitbucket.